Make use of subprocess for running gpg
In order to make GnuPGInterface more portable to systems which do not
support fork/exec, use the subprocess module to run gpg.

This change should be almost completely backwards-compatible and still
passes all unit tests.  The only significant issue is the handling of
file descriptors, which I intend to address in a future commit.

Note:  Required Python version bumped to 2.4 in setup.cfg due to
requirement on subprocess module.

Signed-off-by: Kevin Locke <klocke@digitalenginesoftware.com>