Close unnecessary file descriptors in preexec
In cases where we can't tell subprocess to close all file descriptors,
create a preexec function to close all unnecessary file descriptors.
This function is only required (and supported) on Unix-like systems.

Without this function, pipes which are not explicitly set FD_CLOEXEC may
experience delays in reading EOF due to an unanticipated open end in
gpg.  This will result in a deadlock if the read end of such a pipe is
passed in attach_fds.

Signed-off-by: Kevin Locke <klocke@digitalenginesoftware.com>