Pass OS file handles to gpg on Windows
On the Windows platform, C file descriptors are not inheritable by child
processes.  OS file handles can be inheritable (but are not created as
inheritable by Python, see issue #4708).  To work around these
limitations, get the OS handle for the C file descriptor then duplicate
the file handle as inheritable and pass a string with the handle number
to gpg (which already does the converse).

This implementation uses private modules and methods of those modules.
This leaves the implementation open to future breakage, but there
doesn't appear to be an alternative.

Signed-off-by: Kevin Locke <klocke@digitalenginesoftware.com>