Encode passphrase when required
In Python 3, allow the passphrase attribute to be set to either a string
(for API-compatibility) or a bytes object.  When set to a string, it
must be encoded before being written to the passphrase file descriptor.

Note:  The encoding is only done in Python 3 where we can be sure that
it has not already been encoded.  Otherwise we risk breaking current
users that rely on passphrase not being encoded (or not being
double-encoded if it has already been encoded).

Signed-off-by: Kevin Locke <klocke@digitalenginesoftware.com>