Avoid "except as" syntax
Python 2.5 does not support the "except as" syntax while Python 3 does
not support the "except ," syntax.  In order to support both, get the
error instance from exc_info().

This is a bit less efficient, so it has been marked with a FIXME so that
when 2.5 support is dropped the new syntax can be used.

Signed-off-by: Kevin Locke <klocke@digitalenginesoftware.com>