Big whitespace cleanup
Python 3 is a lot more strict about whitespace and does not allow mixing
of tabs and spaces.  Therefore, replace all tabs with spaces so that the
source file is indented with 4 spaces everywhere.  Also remove trailing
whitespace from lines as it is unnecessary and simply wastes space and
confuses editors.

Signed-off-by: Kevin Locke <klocke@digitalenginesoftware.com>