Increase size of plaintext in test_attach_fhs_pipe
Also exercise pipe buffering by making the plaintext large enough that
the buffer should fill completely.  This should help identify some
blocking issues if any are introduced.

Wrap the pipe reader in a file object so that we do not need to know the
maximum read size ahead of time.

Signed-off-by: Kevin Locke <klocke@digitalenginesoftware.com>