Bump required python version to 2.5
This may not be strictly required, but I don't have Python 2.4 around to
test and it old enough that it is not worth it to me to find and
install/test.

If the module is, or can be made to be, or has a significant use case
for being compatible with Python 2.4, I will consider reverting this
change.

Signed-off-by: Kevin Locke <klocke@digitalenginesoftware.com>