Ensure child descriptors are not set FD_CLOEXEC
For backwards compatibility, set all file descriptors intended for the
child process to remain open after exec.

Arguably passing a descriptor set FD_CLOEXEC into a function where it is
intended to be used by an exec'd process is an argument error.  However,
for backwards compatibility we just clear the flag.

Signed-off-by: Kevin Locke <klocke@digitalenginesoftware.com>